+++ R/links.R
@@ -147,10 +147,16 @@ ConnectionsToLinks <- function(conns, ccans = NULL, threshold = 0) {
 
 #' Link peaks to genes
 #'
-#' Find peaks that are correlated with gene expression. Fits a generalized
-#' linear model with elastic net regularization between peak accessibility and
-#' gene expression for all peaks within a given distance threshold from the gene
-#' TSS.
+#' Find peaks that are correlated with the expression of nearby genes.
+#' For each gene, this function computes the correlation coefficient between
+#' the gene expression and accessibility of each peak within a given distance
+#' from the gene TSS, and computes an expected correlation coefficient for each
+#' peak given the GC content, accessibility, and length of the peak. The expected
+#' coefficent values for the peak are then used to compute a z-score and p-value.
+#'
+#' This function was inspired by the method originally described by SHARE-seq
+#' (Sai Ma et al. 2020, Cell). Please consider citing the original SHARE-seq
+#' work if using this function: \url{https://doi.org/10.1016/j.cell.2020.09.056}
 #'
 #' @param object A Seurat object
 #' @param peak.assay Name of assay containing peak information
@@ -158,30 +164,27 @@ ConnectionsToLinks <- function(conns, ccans = NULL, threshold = 0) {
 #' @param expression.slot Name of slot to pull expression data from
 #' @param gene.coords GRanges object containing coordinates of genes in the
 #' expression assay. If NULL, extract from gene annotations stored in the assay.
-#' @param binary Binarize the peak counts (default FALSE)
-#' @param alpha Alpha parameter for elastic net regularization. Controls balance
-#' between LASSO (alpha=1) and ridge (alpha=0) regularization.
 #' @param distance Distance threshold for peaks to include in regression model
 #' @param min.cells Minimum number of cells positive for the peak and gene
 #' needed to include in the results.
-#' @param expression.threshold Minimum value for a gene to be classified as
-#' expressed. Only used for filtering genes from regression.
-#' @param keep.all Retain all peaks in the output. Useful for knowing what peaks
-#' had 0 coefficient and what were not tested due to expression or accessibility
-#' thresholds.
 #' @param genes.use Genes to test. If NULL, determine from expression assay.
-#' @param family Response type (for gene expression values). Valid values are
-#' "gaussian", "poisson", "binomial", or a \code{\link[stats]{glm}}-family
-#' object
+#' @param method Which correlation coefficient to compute. Can be "pearson"
+#' (default), "spearman", or "kendall".
+#' @param n_sample Number of peaks to sample at random when computing the null
+#' distribution.
+#' @param pvalue_cutoff Minimum p-value required to retain a link. Links with a
+#' p-value equal or greater than this value will be removed from the output.
+#' @param score_cutoff Minimum absolute value correlation coefficient for a link
+#' to be retained
 #' @param verbose Display messages
 #'
 #' @importFrom Seurat GetAssayData
-#' @importFrom glmnet cv.glmnet
-#' @importFrom stats coef
+#' @importFrom stats pnorm sd
 #' @importFrom Matrix sparseMatrix rowSums
 #' @importFrom future.apply future_lapply
 #' @importFrom future nbrOfWorkers
 #' @importFrom pbapply pblapply
+#' @importMethodsFrom Matrix t
 #'
 #' @return Returns a Seurat object with the \code{Links} information set, with
 #' each recorded link being the correlation coefficient between the
@@ -195,14 +198,13 @@ LinkPeaks <- function(
   expression.assay,
   expression.slot = "data",
   gene.coords = NULL,
-  binary = FALSE,
-  alpha = 0.95,
   distance = 5e+05,
   min.cells = 10,
-  expression.threshold = 0.1,
-  family = "poisson",
+  method = "pearson",
   genes.use = NULL,
-  keep.all = FALSE,
+  n_sample = 200,
+  pvalue_cutoff = 0.05,
+  score_cutoff = 0.05,
   verbose = TRUE
 ) {
   if (is.null(x = gene.coords)) {
@@ -210,14 +212,28 @@ LinkPeaks <- function(
       ranges = Annotation(object = object[[peak.assay]])
     )
   }
+  meta.features <- GetAssayData(
+    object = object, assay = peak.assay, slot = "meta.features"
+  )
+  features.match <- c("GC.percent", "count")
+  if (!("GC.percent" %in% colnames(x = meta.features))) {
+    stop("GC content per peak has not been computed.\n",
+         "Run RegionsStats before calling this function.")
+  }
   peak.data <- GetAssayData(
     object = object, assay = peak.assay, slot = 'counts'
   )
+  if (!("count" %in% colnames(x = meta.features))) {
+    # compute total count
+    hvf.info <- FindTopFeatures(object = peak.data)
+    hvf.info <- hvf.info[rownames(x = meta.features), ]
+    meta.features <- cbind(meta.features, hvf.info)
+  }
   expression.data <- GetAssayData(
     object = object, assay = expression.assay, slot = expression.slot
   )
-  peakcounts <- rowSums(x = peak.data > 0)
-  genecounts <- rowSums(x = expression.data > expression.threshold)
+  peakcounts <- meta.features[rownames(x = peak.data), "count"]
+  genecounts <- rowSums(x = expression.data > 0)
   peaks.keep <- peakcounts > min.cells
   genes.keep <- genecounts > min.cells
   peak.data <- peak.data[peaks.keep, ]
@@ -238,9 +254,6 @@ LinkPeaks <- function(
       " peaks"
     )
   }
-  if (binary) {
-    peak.data <- BinarizeCounts(object = peak.data)
-  }
   genes <- rownames(x = expression.data)
   gene.coords.use <- gene.coords[gene.coords$gene_name %in% genes,]
   peaks <- granges(x = object[[peak.assay]])
@@ -251,8 +264,13 @@ LinkPeaks <- function(
     distance = distance
   )
   genes.use <- colnames(x = peak_distance_matrix)
+  all.peaks <- rownames(x = peak.data)
+
+  peak.data <- t(x = peak.data)
+
   coef.vec <- c()
   gene.vec <- c()
+  zscore.vec <- c()
   if (nbrOfWorkers() > 1) {
     mylapply <- future_lapply
   } else {
@@ -265,39 +283,81 @@ LinkPeaks <- function(
     FUN = function(i) {
       peak.use <- as.logical(x = peak_distance_matrix[, genes.use[[i]]])
       gene.expression <- expression.data[genes.use[[i]], ]
+      gene.chrom <- as.character(x = seqnames(x = gene.coords.use[i]))
+
       if (sum(peak.use) < 2) {
         # no peaks close to gene
-        return(list("gene" = NULL, "coef" = NULL))
+        return(list("gene" = NULL, "coef" = NULL, "zscore" = NULL))
       } else {
-        peak.access <- t(x = peak.data[peak.use, ])
-        cvfit <- cv.glmnet(
-          x = peak.access,
-          y = gene.expression,
-          alpha = alpha,
-          family = family
+        peak.access <- peak.data[, peak.use]
+        coef.result <- cor(
+          x = as.matrix(x = peak.access),
+          y = as.matrix(x = gene.expression),
+          method = method
         )
-        lambda <- cvfit$lambda.1se
-        coef.results <- coef(object = cvfit, s = lambda)
-        coef.results <- coef.results[2:nrow(x = coef.results), ]
-        if (keep.all) {
-          # keep all genes so we know what was tested
-          coef.results.filtered <- coef.results
+        coef.result <- coef.result[abs(x = coef.result) > score_cutoff, , drop = FALSE]
+
+        if (nrow(x = coef.result) == 0) {
+          return(list("gene" = NULL, "coef" = NULL, "zscore" = NULL))
         } else {
-          coef.results.filtered <- coef.results[coef.results != 0]
+
+          # select peaks at random with matching GC content and accessibility
+          # sample from peaks on a different chromosome to the gene
+          peaks.test <- rownames(x = coef.result)
+          trans.peaks <- all.peaks[
+            !grepl(pattern = paste0("^", gene.chrom), x = all.peaks)
+          ]
+          meta.use <- meta.features[trans.peaks, ]
+          pk.use <- meta.features[peaks.test, ]
+          bg.peaks <- lapply(
+            X = seq_len(length.out = nrow(x = pk.use)),
+            FUN = function(x) {
+              MatchRegionStats(
+                meta.feature = meta.use,
+                query.feature = pk.use[x, , drop = FALSE],
+                features.match = c("GC.percent", "count", "sequence.length"),
+                n = n_sample,
+                verbose = FALSE
+              )
+            }
+          )
+          # run background correlations
+          bg.access <- peak.data[, unlist(x = bg.peaks)]
+          bg.coef <- cor(
+            x = as.matrix(x = bg.access),
+            y = as.matrix(x = gene.expression),
+            method = method
+          )
+          zscores <- vector(mode = "numeric", length = length(x = peaks.test))
+          for (j in seq_along(along.with = peaks.test)) {
+            coef.use <- bg.coef[(((j - 1) * n_sample) + 1):(j * n_sample), ]
+            z <- (coef.result[j] - mean(x = coef.use)) / sd(x = coef.use)
+            zscores[[j]] <- z
+          }
+          names(x = coef.result) <- peaks.test
+          names(x = zscores) <- peaks.test
+          zscore.vec <- c(zscore.vec, zscores)
+          gene.vec <- c(gene.vec, rep(i, length(x = coef.result)))
+          coef.vec <- c(coef.vec, coef.result)
         }
-        if (length(x = coef.results.filtered) == 0) {
-          return(list("gene" = NULL, "coef" = NULL))
+        gc(verbose = FALSE)
+        pval.vec <- pnorm(q = -abs(x = zscore.vec))
+        links.keep <- pval.vec < pvalue_cutoff
+        if (sum(x = links.keep) == 0) {
+          return(list("gene" = NULL, "coef" = NULL, "zscore" = NULL))
         } else {
-          gene.vec <- c(gene.vec, rep(i, length(x = coef.results.filtered)))
-          coef.vec <- c(coef.vec, coef.results.filtered)
+          gene.vec <- gene.vec[links.keep]
+          coef.vec <- coef.vec[links.keep]
+          zscore.vec <- zscore.vec[links.keep]
+          return(list("gene" = gene.vec, "coef" = coef.vec, "zscore" = zscore.vec))
         }
       }
-      return(list("gene" = gene.vec, "coef" = coef.vec))
     }
   )
   # combine results
   gene.vec <- do.call(what = c, args = sapply(X = res, FUN = `[[`, 1))
   coef.vec <- do.call(what = c, args = sapply(X = res, FUN = `[[`, 2))
+  zscore.vec <- do.call(what = c, args = sapply(X = res, FUN = `[[`, 3))
 
   if (length(x = coef.vec) == 0) {
     if (verbose) {
@@ -317,9 +377,20 @@ LinkPeaks <- function(
   )
   rownames(x = coef.matrix) <- genes.use
   colnames(x = coef.matrix) <- names(x = peak.key)
-
-  # create links granges and add to assay
   links <- LinksToGRanges(linkmat = coef.matrix, gene.coords = gene.coords.use)
+  # add zscores
+  z.matrix <- sparseMatrix(
+    i = gene.vec,
+    j = peak.key[names(x = zscore.vec)],
+    x = zscore.vec,
+    dims = c(length(x = genes.use), max(peak.key))
+  )
+  rownames(x = z.matrix) <- genes.use
+  colnames(x = z.matrix) <- names(x = peak.key)
+  z.lnk <- LinksToGRanges(linkmat = z.matrix, gene.coords = gene.coords.use)
+  links$zscore <- z.lnk$score
+  links$pvalue <- pnorm(q = -abs(x = links$zscore))
+  links <- links[links$pvalue < pvalue_cutoff]
   Links(object = object[[peak.assay]]) <- links
   return(object)
 }