+++ R/utilities.R
@@ -828,8 +828,20 @@ LookupGeneCoords <- function(object, gene, assay = NULL) {
 #' query regions for any given set of characteristics, specified in the input
 #' \code{meta.feature} dataframe.
 #'
-#' @param meta.feature A dataframe containing DNA sequence information
-#' @param regions Set of query regions. Must be present in rownames.
+#' For each requested feature to match, a density
+#' distribution is estimated using the \code{\link[stats]{density}} function,
+#' and a set of weights for each feature in the dataset estimated based on the
+#' density distribution. If multiple features are to be matched (for example,
+#' GC content and overall accessibility), a joint density distribution is then
+#' computed by multiplying the individual feature weights. A set of features
+#' with characteristics matching the query regions is then selected using the
+#' \code{\link[base]{sample}} function, with the probability of randomly
+#' selecting each feature equal to the joint density distribution weight.
+#'
+#' @param meta.feature A dataframe containing DNA sequence information for
+#' features to choose from
+#' @param query.feature A dataframe containing DNA sequence information for
+#' features to match.
 #' @param n Number of regions to select, with characteristics matching the query
 #' @param features.match Which features of the query to match when selecting a
 #' set of regions. A vector of column names present in the feature metadata can
@@ -846,15 +858,17 @@ LookupGeneCoords <- function(object, gene, assay = NULL) {
 #' metafeatures <- Seurat::GetAssayData(
 #'   object = atac_small[['peaks']], slot = 'meta.features'
 #' )
+#' query.feature <- metafeatures[1:10, ]
+#' features.choose <- metafeatures[11:nrow(metafeatures), ]
 #' MatchRegionStats(
-#'   meta.feature = metafeatures,
-#'   regions = head(rownames(metafeatures), 10),
+#'   meta.feature = features.choose,
+#'   query.feature = query.feature,
 #'   features.match = "percentile",
 #'   n = 10
 #' )
 MatchRegionStats <- function(
   meta.feature,
-  regions,
+  query.feature,
   features.match = c("GC.percent"),
   n = 10000,
   verbose = TRUE,
@@ -863,27 +877,21 @@ MatchRegionStats <- function(
   if (!inherits(x = meta.feature, what = 'data.frame')) {
     stop("meta.feature should be a data.frame")
   }
+  if (!inherits(x = query.feature, what = "data.frame")) {
+    stop("query.feature should be a data.frame")
+  }
   if (length(x = features.match) == 0) {
     stop("Must supply at least one sequence characteristic to match")
   }
-  if (!(all(regions %in% rownames(x = meta.feature)))) {
-    warning("Not all regions are present in meta.features. ",
-            "Removing missing regions.")
-    regions <- intersect(x = regions, y = rownames(x = meta.feature))
-  }
-  mf.query <- meta.feature[regions, ]
-  choosefrom <- setdiff(
-    x = rownames(x = meta.feature), y = rownames(x = mf.query)
-  )
-  if (length(x = choosefrom) < n) {
-    n <- length(x = choosefrom)
+  if (nrow(x = meta.feature) < n) {
+    n <- nrow(x = meta.feature)
     warning("Requested more features than present in supplied data.
             Returning ", n, " features")
   }
-  features.choose <- meta.feature[choosefrom, ]
-  feature.weights <- rep(0, nrow(features.choose))
-  for (i in features.match) {
-    if (!(i %in% colnames(x = mf.query))) {
+  # features.choose <- meta.feature[choosefrom, ]
+  for (i in seq_along(along.with = features.match)) {
+    featmatch <- features.match[[i]]
+    if (!(featmatch %in% colnames(x = query.feature))) {
       if (i == "GC.percent") {
         stop("GC.percent not present in meta.features.",
              " Run RegionStats to compute GC.percent for each feature.")
@@ -892,23 +900,30 @@ MatchRegionStats <- function(
       }
     }
     if (verbose) {
-      message("Matching ", i, " distribution")
+      message("Matching ", featmatch, " distribution")
     }
-    density.estimate <- density(x = mf.query[[i]], kernel = "gaussian", bw = 1)
+    density.estimate <- density(
+      x = query.feature[[featmatch]], kernel = "gaussian", bw = 1
+    )
     weights <- approx(
       x = density.estimate$x,
       y = density.estimate$y,
-      xout = features.choose[[i]],
+      xout = meta.feature[[featmatch]],
       yright = 0.0001,
       yleft = 0.0001
     )$y
-    feature.weights <- feature.weights + weights
+    if (i > 1) {
+      feature.weights <- feature.weights * weights
+    } else {
+      feature.weights <- weights
+    }
   }
-  feature.select <- sample(
-    x = rownames(x = features.choose),
+  feature.select <- sample.int(
+    n = nrow(x = meta.feature),
     size = n,
     prob = feature.weights
   )
+  feature.select <- rownames(x = meta.feature)[feature.select]
   return(feature.select)
 }