+++ NEWS.md
@@ -24,6 +24,10 @@ and [#260](https://github.com/timoast/signac/issues/260))
 * Removed `qvalue` calculation from `FindMotifs()` ([#223](https://github.com/timoast/signac/issues/223))
 * Fixed bug in `SetAssayData()` when setting the `scale.data` slot
 
+Other changes:
+
+* Improved feature matching in `MatchRegionStats()` function when matching distribution of multiple features (eg, GC content and overall accessibility)
+* Changed parameter names in `MatchRegionStats()`
 
 # Signac 1.0.0